Move specific extraction methods into an helpers namespace
